Update studying.md
The whole "Display order" section is a little bit outdated, it seems that it mixes concepts from the three schedulers, and I personally find it kind of confusing, and I'm not familiar enough with previous schedulers to delete any info. 
I've reorganized the section a little bit, but maybe some extra update / simplification of the whole section would be in order.