document TTS platform discrepancy and fix (#127)
* document TTS platform discrepancy and fix

AnkiDroid renders {{tts FieldName}} as text while other platforms render a (re)play TTS button. We should hide this text if we're on Android. To give AnkiDroid users a (re)play TTS button, we can make one that calls the JS API, and hide this button for all other platforms.

* remove some debug text mistakenly left behind